feat: refactor the RemoteCommunicationPostMessageStream class
